app/requirements.txt
# Web app dependencies
Flask
Werkzeug
# PPTX converter dependencies
python-pptx